cleaning on-start-app middleware / rename post-progression api action
