+++ packages/@coorpacademy-app-review/src/actions/api/fetch-skills.ts
@@ -1,5 +1,5 @@
 import buildTask from '@coorpacademy/redux-task';
-import {Skill} from '../../types/skills';
+import {Skill} from '../../types/common';
 
 export const SKILLS_FETCH_REQUEST = '@@skills/FETCH_REQUEST' as const;
 export const SKILLS_FETCH_SUCCESS = '@@skills/FETCH_SUCCESS' as const;
@@ -10,7 +10,8 @@ export type ReceivedSkills = {
   payload: Skill[];
 };
 
-export const fetchSkills = (token: string) => buildTask({
+export const fetchSkills = (token: string) =>
+  buildTask({
     types: [SKILLS_FETCH_REQUEST, SKILLS_FETCH_SUCCESS, SKILLS_FETCH_FAILURE],
     task: (dispacth, getState, {services}) => services.fetchSkills(token)
   });
