--- packages/@coorpacademy-app-review/src/types/skills.ts
@@ -1,6 +1,0 @@
-export type Skill = {
-  skillRef: string;
-  slidesToReview: number;
-  custom: boolean;
-  name: string;
-};
